# Conflicts:
#	Classes/KeyboardManager.cs
#	Classes/ModifiableHotkey.cs
#	Filesystem/Selector/FileSystemSelector.DragDrop.cs
#	Filesystem/Selector/FileSystemSelector.Draw.cs
#	ItemSelector.cs
#	OtterGuiInternal
#	Raii/EndObjects.cs
#	Text/EndObjects/CollapsingHeader.cs
#	Text/EndObjects/Columns.cs
#	Text/EndObjects/Combo.cs
#	Text/EndObjects/ListBox.cs
#	Text/EndObjects/Menu.cs
#	Text/EndObjects/PopupContextItem.cs
#	Text/EndObjects/TabBar.cs
#	Text/EndObjects/TabItem.cs
#	Text/EndObjects/Table.cs
#	Text/ImUtf8.Clipboard.cs
#	Text/ImUtf8.RotatedText.cs
#	Util.cs
#	Widgets/HexViewer.cs

+++ Widgets/ImGuiResizingTextInput.cs
@@ -32,7 +32,7 @@ public static partial class Widget
             width = ImGui.CalcTextSize(input).X + 10;
 
         ImGui.SetNextItemWidth(width);
-        var ret = ImGui.InputText(label, ref input, maxLength, ImGuiInputTextFlags.EnterReturnsTrue);
+        var ret = ImGui.InputText(label, ref input, (int)maxLength, ImGuiInputTextFlags.EnterReturnsTrue);
         TextInputWidths[id] = ImGui.CalcTextSize(input).X + 10;
         return (ret, id);
     }
